Updating system plugins for 0.3.0
