+++ etc/telegraf.conf
@@ -97,8 +97,8 @@
   # Mountpoints=["/"]
 
 # Read metrics about disk IO by device
-[[plugins.io]]
-  # By default, telegraf will gather stats for all devices including 
+[[plugins.diskio]]
+  # By default, telegraf will gather stats for all devices including
   # disk partitions.
   # Setting devices will restrict the stats to the specified devices.
   # Devices=["sda","sdb"]
