+++ plugins/system/docker.go
@@ -36,44 +36,47 @@ func (s *DockerStats) Gather(acc plugins.Accumulator) error {
 
 		cts := cont.CPU
 
-		acc.Add("user", cts.User, tags)
-		acc.Add("system", cts.System, tags)
-		acc.Add("idle", cts.Idle, tags)
-		acc.Add("nice", cts.Nice, tags)
-		acc.Add("iowait", cts.Iowait, tags)
-		acc.Add("irq", cts.Irq, tags)
-		acc.Add("softirq", cts.Softirq, tags)
-		acc.Add("steal", cts.Steal, tags)
-		acc.Add("guest", cts.Guest, tags)
-		acc.Add("guest_nice", cts.GuestNice, tags)
+		fields := map[string]interface{}{
+			"user":       cts.User,
+			"system":     cts.System,
+			"idle":       cts.Idle,
+			"nice":       cts.Nice,
+			"iowait":     cts.Iowait,
+			"irq":        cts.Irq,
+			"softirq":    cts.Softirq,
+			"steal":      cts.Steal,
+			"guest":      cts.Guest,
+			"guest_nice": cts.GuestNice,
 
-		acc.Add("cache", cont.Mem.Cache, tags)
-		acc.Add("rss", cont.Mem.RSS, tags)
-		acc.Add("rss_huge", cont.Mem.RSSHuge, tags)
-		acc.Add("mapped_file", cont.Mem.MappedFile, tags)
-		acc.Add("swap_in", cont.Mem.Pgpgin, tags)
-		acc.Add("swap_out", cont.Mem.Pgpgout, tags)
-		acc.Add("page_fault", cont.Mem.Pgfault, tags)
-		acc.Add("page_major_fault", cont.Mem.Pgmajfault, tags)
-		acc.Add("inactive_anon", cont.Mem.InactiveAnon, tags)
-		acc.Add("active_anon", cont.Mem.ActiveAnon, tags)
-		acc.Add("inactive_file", cont.Mem.InactiveFile, tags)
-		acc.Add("active_file", cont.Mem.ActiveFile, tags)
-		acc.Add("unevictable", cont.Mem.Unevictable, tags)
-		acc.Add("memory_limit", cont.Mem.HierarchicalMemoryLimit, tags)
-		acc.Add("total_cache", cont.Mem.TotalCache, tags)
-		acc.Add("total_rss", cont.Mem.TotalRSS, tags)
-		acc.Add("total_rss_huge", cont.Mem.TotalRSSHuge, tags)
-		acc.Add("total_mapped_file", cont.Mem.TotalMappedFile, tags)
-		acc.Add("total_swap_in", cont.Mem.TotalPgpgIn, tags)
-		acc.Add("total_swap_out", cont.Mem.TotalPgpgOut, tags)
-		acc.Add("total_page_fault", cont.Mem.TotalPgFault, tags)
-		acc.Add("total_page_major_fault", cont.Mem.TotalPgMajFault, tags)
-		acc.Add("total_inactive_anon", cont.Mem.TotalInactiveAnon, tags)
-		acc.Add("total_active_anon", cont.Mem.TotalActiveAnon, tags)
-		acc.Add("total_inactive_file", cont.Mem.TotalInactiveFile, tags)
-		acc.Add("total_active_file", cont.Mem.TotalActiveFile, tags)
-		acc.Add("total_unevictable", cont.Mem.TotalUnevictable, tags)
+			"cache":                  cont.Mem.Cache,
+			"rss":                    cont.Mem.RSS,
+			"rss_huge":               cont.Mem.RSSHuge,
+			"mapped_file":            cont.Mem.MappedFile,
+			"swap_in":                cont.Mem.Pgpgin,
+			"swap_out":               cont.Mem.Pgpgout,
+			"page_fault":             cont.Mem.Pgfault,
+			"page_major_fault":       cont.Mem.Pgmajfault,
+			"inactive_anon":          cont.Mem.InactiveAnon,
+			"active_anon":            cont.Mem.ActiveAnon,
+			"inactive_file":          cont.Mem.InactiveFile,
+			"active_file":            cont.Mem.ActiveFile,
+			"unevictable":            cont.Mem.Unevictable,
+			"memory_limit":           cont.Mem.HierarchicalMemoryLimit,
+			"total_cache":            cont.Mem.TotalCache,
+			"total_rss":              cont.Mem.TotalRSS,
+			"total_rss_huge":         cont.Mem.TotalRSSHuge,
+			"total_mapped_file":      cont.Mem.TotalMappedFile,
+			"total_swap_in":          cont.Mem.TotalPgpgIn,
+			"total_swap_out":         cont.Mem.TotalPgpgOut,
+			"total_page_fault":       cont.Mem.TotalPgFault,
+			"total_page_major_fault": cont.Mem.TotalPgMajFault,
+			"total_inactive_anon":    cont.Mem.TotalInactiveAnon,
+			"total_active_anon":      cont.Mem.TotalActiveAnon,
+			"total_inactive_file":    cont.Mem.TotalInactiveFile,
+			"total_active_file":      cont.Mem.TotalActiveFile,
+			"total_unevictable":      cont.Mem.TotalUnevictable,
+		}
+		acc.AddFields("docker", fields, tags)
 	}
 
 	return nil
